Move reviewer to ts/reviewer
